Consolidate files, components, settings, imports
:cleanUp:
