+++ client/src/components/JamRoom.jsx
@@ -1,6 +1,4 @@
 import React, { Component } from 'react';
-import UserInstrument from './UserInstrument';
-import FriendInstrument from './FriendInstrument';
 import Piano from './Piano';
 import Drums from './Drums';
