--- client/src/containers/FriendInstrument.jsx
@@ -1,14 +1,0 @@
-import React, { Component } from 'react';
-
-class FriendInstrument extends Component {
-
-  render() {
-    return (
-      <div className="friendInstrument">
-        <h1>Invite a friend to play!</h1>
-      </div>
-    );
-  }
-}
-
-export default FriendInstrument;
