--- client/src/containers/Text.jsx
@@ -1,13 +1,0 @@
-import React, { Component } from 'react';
-
-
-class Text extends Component {
-  render() {
-    return (
-      <div>
-      </div>
-    );
-  }
-}
-
-export default Text;
