+++ client/src/index.jsx
@@ -3,15 +3,13 @@ import { render } from 'react-dom';
 import { Router, Route, IndexRoute, browserHistory } from 'react-router';
 import injectTapEventPlugin from 'react-tap-event-plugin';
 
-import App from './containers/App';
+import App from './components/App';
 import LandingPage from './components/LandingPage';
-import Login from './containers/Login';
-import Signup from './containers/Signup';
-import SelectInstrument from './containers/SelectInstrument';
-import SelectRoom from './containers/SelectRoom';
-import JamRoom from './containers/JamRoom';
-import Room from 'Room';
-import Invalid from 'Invalid';
+import Login from './components/Login';
+import Signup from './components/Signup';
+import JamRoom from './components/JamRoom';
+import Room from './components/Room';
+import Invalid from './components/Invalid';
 import Metronome from './components/Metronome';
 
 // Needed for onTouchTap
@@ -24,9 +22,6 @@ render((
       <IndexRoute component={LandingPage} />
       <Route path="login" component={Login} />
       <Route path="signup" component={Signup} />
-      <Route path="selectInstrument" component={SelectInstrument} />
-      <Route path="selectRoom" component={SelectRoom} />
-      <Route path="jam" component={JamRoom} />
       <Route path="room/:roomId" component={Room} />
       <Route path="metronome" component={Metronome} />
       <Route path="*" component={Invalid} />
