+++ client/src/instruments/store.js
@@ -1,8 +1,5 @@
-// import piano
-// import fry
-import fry from 'fry';
-import piano from 'piano';
-// import piano from ''
+import fry from './fry';
+import piano from './piano';
 
 const store = {
   fry: fry.playNote,
